// types/person.d.ts
// types/person.d.ts

export interface Person {
    name: string;
    height: string;
    mass: string;
    hair_color: string;
    skin_color: string;
    eye_color: string;
    birth_year: string;
    gender: string;
    homeworld: string;
    films: string[],
    starships: string[],
    vehicles: string[],
    species: string[],
    url: string;
  }
  